Upload the benchmarks in a separate GitHub step
