+++ .github/workflows/run-benchmarks.yml
@@ -24,9 +24,15 @@ jobs:
           ./gradlew -p benchmark :app:assembleRelease :microbenchmark:assembleReleaseAndroidTest
       # Step can be removed if/when gcloud is added to the macos image - See https://github.com/actions/virtual-environments/issues/4639
       - uses: google-github-actions/setup-gcloud@98ddc00a17442e89a24bbf282954a3b65ce6d200 #v2.1.0
-      - run: ./scripts/run-benchmarks.main.kts
+      - run: |
+          # Run the micro benchmarks and write the results to benchmark/build/benchmarks.json
+          ./scripts/run-benchmarks.main.kts
         env:
           GOOGLE_SERVICES_JSON: ${{ secrets.APOLLO_KOTLIN_GOOGLE_SERVICES_JSON }}
+      - run: |
+          # Upload the results to DataDog and update the benchmarks issue
+          ./scripts/upload-benchmarks.main.kts
+        env:
           DD_API_KEY: ${{ secrets.DD_API_KEY }}
           GITHUB_TOKEN: ${{ github.token }}
       - run: |
